fix: fix NotEq case in anticipation of non-literals
The code as written fails when the "other" [min, max] interval has size greater than 1.

Call the left-hand-side's interval: `[lmin, lmax]` and the right-hand-side's interval: `[rmin,
rmax]`. The current expression is:

    (lmin == rmin) && (rmin == lmax)
    ||
    (lmin == rmax) && (rmax == lmax)

By transitivity, in either conjunction, lmin == lmax. That means the column has a single known
value: lmin (equivalently: lmax).

If the right-hand-side is only ever literals or expressions thereof, its interval is always `[x,
x]`. In that case, this expression is tantamount to: is the column constant and equal to `x`?

If the right-hand-side is either another column or some non-deterministic expression, the interval
could be, for example: `[10, 11]`. If the column is known to be the constant value `10` (i.e. its
min and max are 10), we cannot prune this chunk! The rows where the right-hand-side is 11 would
satisfy the inequality.

I'm also open to: `(lmin == lmax) && (lmin == rmin)` with a comment indicating that, because we only
have literals, rmin is necessarily equal to rmax and thus rmin _is_ the, known, right-hand-side
value.