fix: just skip the universal2 wheel entirely
> You need to just stop.
> Like can you just not

We already produce a wheel for macOS arm64 and macOS x86_64. Pip
will pick the correct wheel. The only real benefit of a universal2
wheel is that, in theory, if there was a lot of shared resources
between the amr64 and x86_64 builds, we would only need to upload
that to PyPI once. In our case, most of the bulk seems to come from
the binary anyway (75MiB before compression).

I suppose another benefit is for folks not using pip: they need not
determine which wheel applies. :shrug:.