fix: binary_numeric is now correct (+ tests!)
Lessons learned: do not merge without at least basic tests.

We check for a constant array but do not create a new constant array
with the appropriate length for the child on which we want to delegate
the binary numeric operation. That is now fixed.

I also added a very basic test suite for applying the (now six)
numeric operations on pairs of arrays where one of the two
are constant.

In order to properly support constant arrays on the left-
hand-side, I added FlippedSub and FlippedDiv which allow
us to commute/flip the operator so that we need not teach
ConstantArray to check if its right-hand-side supports
an operation on constants.