feat: compare array metadata against checked in goldenfiles
This is part of a general epic (is that word taboo?) around stability. I have
a separate changeset that stores an entire Vortex file and checks for round-trip
compatibility.

However, it's also helpful to know on a per-encoding level that the metadata format
is stable.

This PR adds [goldenfile](https://crates.io/crates/goldenfile) tests for all
encoding metadata. This is a bit simplistic: we don't check for backward/forward
compatibility, this is just an attempt to make it clearer on PRs when changes are
made to the metadata format.

Now when changes are made to Array metadata, you should re-run the test with

```
UPDATE_GOLDENFILES=1 cargo test -p ...
```

and that will update the checked-in metadatas.