feat: teach BitPackedEncoding canonical_into
Runtime for a 100k element array is reduced to ~60% of `into_canonical`.

This runtime reduction only holds when the builder avoids all allocations! The builder must be large
enough to contain both the entire decoded array and also any scratch space needed by the
decoder. For BitPacked, that means it needs 1023 empty element spots in case it needs to decode a
1024 element chunk to retrieve just one element.

```
canonicalize_bench    fastest       │ slowest       │ median        │ mean          │ samples │ iters
├─ test               43.24 µs      │ 221.6 µs      │ 47.62 µs      │ 47.49 µs      │ 100000  │ 100000
├─ canonical_into                   │               │               │               │         │
│  ╰─ u32                           │               │               │               │         │
│     ╰─ (100000, 3)  24.12 µs      │ 104.7 µs      │ 24.29 µs      │ 24.39 µs      │ 100000  │ 100000
╰─ into_canonical                   │               │               │               │         │
   ╰─ u32                           │               │               │               │         │
      ╰─ (100000, 3)  39.54 µs      │ 249.6 µs      │ 39.91 µs      │ 40.95 µs      │ 100000  │ 100000
```