feat: GeoVortex
This branch serves as an implementation sketch for a new series of extension types around geospatial vector data.

It provides new extension type IDs, as well as bindings to GeoArrow so that data can be converted back and forth into the canonical format for sharing with tools that speak GeoArrow.

This is just a sketch and is meant to inform some ideas around what's missing with extension types