feat: Make `Identifier` an enum (#3478)
Using an empty string as a sentinel value for the "identity" identifier
seems bug-prune, and the API isn't very ergonomic.

The naming in this PR is a bit atrocious, but I'm more than open to
suggestions.

Also added some comments that made sense to me but might not be 100%
correct or clear.

---------

Signed-off-by: Adam Gutglick <adam@spiraldb.com>