feat: link duckdb dynamically (#3506)
When linking static libraries, the linker includes the entire object
file, not just the symbols needed. The static library might not be self
contained and can reference symbols that require additional libraries to
be linked, as it was the case here with `libduckdb_static.a`.

On macOS, the static linking suceeded in the previous setup already
because of the linker flags we set in `.cargo/config` for which there is
no identical equivalent for Linux.

```
[target.aarch64-apple-darwin]
rustflags = ["-C", "link-arg=-undefined", "-C", "link-arg=dynamic_lookup"]
```

Signed-off-by: Alexander Droste <alexander.droste@protonmail.com>