fix: DecimalByteParts compare with overflow and nullable elements (#3789)
Fuzzer is [currently
failing](https://github.com/vortex-data/vortex/actions/runs/16105049077/job/45439258523)
on an optimization pathway in DecimalBytePartsArray, where the constant
cannot be represented in the narrowed PType of DBP's narrowed PType.

So for example if you have a nullable DecimalByteParts with values
`None, Some(1i32), Some(2i32)` and you do the comparison `<=
ConstantArray(Some(999999999999999i128))`, we'd first try and cast
`999999999999999i128` to an `i32`, and when that fails we'd assume that
it is larger than all values in the original array, so immediately
return `ConstantArray(true)`.

However, that disregards the validity information. I've updated the
branch so that the optimized pathway is only taken if the array and the
constant are non-null

Separately I removed the references to lower_parts, it was a lot of
cruft and idk when we're going to implement it so just put a note that
we're receiving a field for it in the struct and metadata.

---------

Signed-off-by: Andrew Duffy <andrew@a10y.dev>