feat: Add optimize ArrayOp with VBView implementation (#3825)
Adds a new OperationVTable method `optimize` which will return a new
array with unused data garbage-collected. This is similar to the
arrow-rs `StringViewArray::gc` method. See doc comments in the PR for
more info.

This implements a simple optimize operation that will fully rebuild a
VarBinViewArray if there is any wasted space in the buffers.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>