feat: call optimize in compressor
For each DType compressor, we have a fallback `UncompressedScheme`
for when no real compression strategies are applicable. Currently
this just returns the original array, this PR just adds a call
to the new `optimize()` arrayop before returning, so that even if
we don't compress at least we optimize the result. Currently this
is only applicable for VarBinView arrays.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>