chore[docs]: deny missing_docs on vortex-scalar, vortex-utils, & vortex-flatbuffers (#3907)
I gave the following prompt to Claude Code:

> Our goal is to write high quality docstrings in idiomatic Rust style.
We would like to enforce this for certain crates, by applying the
`#![deny(missing_docs)]` directive to those crates. Please apply that
directive and then write docstrings for all of the public methods and
structs in the indicated crates. For now, let's start by working on the
vortex-scalar, vortex-utils, and vortex-flatbuffers crates.

I then skimmed the changes in GH, made some minor revisions with help
from Cursor. AI is taking our jobs?

---------

Signed-off-by: Will Manning <will@willmanning.io>