Vortex - C++ API (#3792)
This PR uses the cxx crate to bridge Rust and C++. The trade-offs
between cxx and directly wrapping the vortex-ffi C API is roughly the
follows:

Pros of cxx:
1. Safety is guaranteed by cxx, no need to directly use the unsafe C API
(not all cases though). In a sense that the Rust-side code is easier to
write and maintain.
2. Better ergonomic by providing Rust data structure
[bindings](https://cxx.rs/bindings.html) with RAII support in C++ like
`rust::Box` and `rust::Str` .

Cons:
1. Less flexibility because we cannot control what FFI is generated, and
we may lack some type bindings support across the FFI.
2. May eventually have duplicate code as `vortex-ffi`

I also tried to find best practices in other repos and seems
[opendal](https://github.com/apache/opendal/tree/main/bindings) uses cxx
for C++ and cbindgen for C with separate code paths, too. I think we can
follow the same approach at least for now for a limited C++ API (e.g.,
to/from Arrow). We can refactor if we need direct-access to C FFI in the
future.

*The first PR intends to setup a C++ directory structure and convention
instead of fixing to use cxx.*

cc @royguo

---------

Signed-off-by: Xinyu Zeng <xinyuzeng@tencent.com>
Co-authored-by: Robert Kruszewski <github@robertk.io>