chore: fix type bound on Scalar::binary
There isn't really that much that implements
Into<Arc<ByteBuffer>>. Instead, we just accept Into<ByteBuffer>
and then Arc it before constructing the value. The internal Arc is
just there to cut down on the size of InnerScalarValue, there aren't any
code paths where we actually get an Arc<ByteBuffer> out of something and
want to turn it into a scalar directly.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>