chore: add PyO3 environment variables (#4350)
This should make builds a bit more consistent.

I also believe it prevents rebuilding some things. If you change
terminals or if something related to the python installation changes,
PyO3 will rebuild everything that is dependent on it. Not sure how
important that is in practice, but this seems like a good thing to have.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>