fix[vortex-array]: propagate preferred type conversion to nested fields
When arrow_type is None, vortex to arrow conversion should respect conversion
to the encoding's preferred arrow type for all nested fields. Previous to this
commit, this preference was lost for any fields within structs or lists.

For example, a struct with a utf8 dictionary field would lose the dictionary
encoding.

Signed-off-by: Alfonso Subiotto Marques <alfonso.subiotto@polarsignals.com>