Feature: Add `is_subset_of` of `is_superset_of` and validate `extend_from_array` (#4445)
This will be helpful in my refactoring of the `builders` module in
`vortex-array` (in draft right now:
https://github.com/vortex-data/vortex/pull/4440)

Note that `extend_from_array` is called in many non-test places yet is
not validated at all. Given that this is a public method, this is a bug
waiting to happen...

I have an interesting thought though: should a `DType::Primitive(U64,
nullable)` be a superset of `DType::Primitive(U16, non-nullable)`?

In theory, we should be able to extend an array builder of nullable
`U64` with an array of non-nullable `U8`.

I'll leave that to a separate PR since this idea might have some other
implications.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>