Dispatch each I/O request (#4464)
An attempt at short-term fix for
https://github.com/vortex-data/vortex/issues/4400

The coalescing driver was spawned as a single task, and so landed on a
single thread of the I/O dispatcher regardless of how many threads you
gave it. This PR spawns each read request to allow other threads a
chance to play ball.

The reason the Polars patch doesn't work is that we need to continue to
dispatch work in order to provide a Tokio runtime within non-Tokio
clients, such as DuckDB. This is a bigger change as part of #4406