make canonical array constructors consistent
Adds `new_unchecked` to all arrays and cleans up safety documentation
and validation in general.

Also renames the `BoolArray::new` method to
`BoolArray::new_from_bool_buffer` to better reflect what is actually
happening.

Note that many of the arrays in `encodings/` are remaing the same, but
we will leave that for a later PR.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>