fix: teach vortex-duckdb to safely copy validity (#4534)
First attempt here: https://github.com/vortex-data/vortex/pull/4528.

I suspect the reason the API was designed to return BitSlice rather than
`&mut [u8]` is that there's an ambiguity as to which bit is bit zero.

I doubt this change affects the benchmark results reported in #4528.
This implementation is probably ~140% the runtime of the version that
uses u64 copies where possible.

Signed-off-by: Daniel King <dan@spiraldb.com>