feature: Bloom Filters in ZoneMaps
This PR (WIP) augments our existing ZoneMapLayout to also optionally
write bloom filters.

Currently, bloom filters will only be written for UTF-8 columns, the
thought being that arbitrary binary blobs aren't obviously splittable
and might be quite large in practice, so it's better to avoid filtering
this. This is weakly held belief and can revisit later.

This is backwards-compatible, the metadata includes an additional
optional protobuf field and if that field is present, it will point at a
segment ID that contains the bloom filters portion of a ZoneMap.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>