Refactor VortexWrite to return Stream<Buffer> and Footer (#4557)
I attempted this in #4549 , but it was hanging, which has been resolved
by switching over to using the new runtimes API for execution. I could
have made this change separately, but both this PR and switching to
runtimes adds an argument to write_stream function so would have been a
conflict nightmare 😢

---

This revives an old branch I had worked on with @onursatici about
changing the Vortex write API to return a Stream and be completely
independent from the actual file sink.

This PR exposes a new write_stream function, and we can remove the old
write(VortexWrite, ArrayStream) APIs in a future PR. We should decide
whether we want to make that change prior to releasing to avoid multiple
user-facing breaks to this Rust API.

In order to make this change easier, we now pass an explicit eof-of-file
pointer into the writers that they can use to create new streams that
occur "just before the end of the file", as is required for zone maps
for example.

We also remove the custom transpose_stream implementation that was quite
dependent on the polling behaviour of `try_join_all`, which alarmingly
changes when the number of futures reaches 40.

Closes #4549 
Fixes #4552 
Fixes #4064

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>