chore: add r/place and airquality datasets to benchmark
I did not add them to CI because I do not want to incur egress charges constantly
for the pcodec project. Probably the right thing to do is to download all these
datasets and upload them, as Parquet, to an S3 bucket we control. None of the
other datasets currently use authenticated downloads and I did not want to figure
that out today.

Signed-off-by: Daniel King <dan@spiraldb.com>