feat: FoRArray is sorted if its compressed form is sorted
I think this is true right? We previously had `[a, b, c, d, ...]` and we now have `[a - min, b -
min, c - min, d - min, ...]`. If we know `a - min >= 0` (which is true by definition of min), then
sortedness should be preserved because `a - k > b - k` implies `a > b`.

Signed-off-by: Daniel King <dan@spiraldb.com>