Remove local path optimization for object store (#4673)
Closes #4672

This optimization is a bit weird. If the user explicitly asks for object
store, and has configured it with telemetry etc, then it's odd for us to
then ignore that request and use our own I/O layer.

Also assuming an absolute path from root was too naive.

Users are encouraged to either `open_object_store`, or `open(&Path)`. If
we had our own registry of object stores, we could also support `impl
IntoReadSource for Url`

Signed-off-by: Nicholas Gates <nick@nickgates.com>