Operator Updates (#4725)
This PR changes the API a little for operators as well as creates an
executor that performs CSE and pipeline conversion.

It breaks many of the old operator tests and benchmarks, many of which
are commented out. But to avoid long-lived feature branch @joseph-isaacs
and I will continue to merge onto develop.

Execution via operators is gated on a runtime environment variable for
local development until we get this prototype fleshed out and
performant!

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>