Initial version of Masked encoding (#4785)
- Closes
https://github.com/vortex-data/vortex/issues/4784#issuecomment-3347186341

Introduces a new `MaskedArray` encoding, which is an array that holds a
child with no null values and an external validity array, allowing
validity-only operations to be very cheap.

The greatest benefit is that together with constant array, it enables us
to compress data more aggressively and keep it compact for longer.

---------

Signed-off-by: Adam Gutglick <adam@spiraldb.com>