ci: add helpful remarks to filter benchmark results
The top-line summaries on PR benchmarks are very helpful, but we could
still use more help to triage results.

For things like TPC-DS, we get back a ~500 element table that you need
to visually scan to find individual queries with signficant changes.

Here we add a 🚀 emoji for queries that improved by >= 30%,
and a 🚨 for ones that decreased performance by >= 30%.

Signed-off-by: Andrew Duffy <andrew@a10y.dev>