Feature: add listview rebuild + cleanup (#4819)
Adds a `rebuild` method on `ListViewArray` that allows us to rebuild the
array to specific shapes that we want (via a `ListViewRebuildMode`
enum).

This will make it easier to implement an optimization (that we are
tentatively calling `IsZeroCopyToList`) that tells us the
`ListViewArray` has no gaps nor overlaps from the views into the
`elements` buffer, and that the views are all sorted and in order.

I also added an unimplemented `OverlapCompression` rebuild mode if we
ever want to do that.

Also does some cleanup of other `ListView` functionality.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>