Expand expression pushdown to DataFusion (#4398)
Started as a demo for @danking to show how to add more pushdown to
Datafusion, now it includes some more expression (`IS NULL`, `IS NOT
NULL` and `IN LIST`), which only leaves `substr` and dynamic
expressions.
It also allows for pushing more binary operators for arithmetic
operations.

`vortex-compact` seems really noise, but clickbench still shows nice
improvements in a bunch of queries for the btrblocks compression.

---------

Signed-off-by: Adam Gutglick <adam@spiraldb.com>