Chore: `vortex-array` modules cleanup (#4818)
The way in which the modules are structured under
`vortex-array/src/arrays` is quite inconsistent. This PR ensures that
each kind of array consistently has the same structure, which is:

- `array.rs`: contains the array struct definition and implementation
- `compute`: contains any compute function implementations that the
encoding implements in separate files
- `vtable`: contains the array vtable implementations, each in a
separate file. The vtable itself is located in `vtable/mod.rs`

Probably the most controversial thing here is the addition of a `vtable`
module, where each helper trait implementation is in its own file. I
agree that it isn't ideal as it means we have a lot of files named the
same thing, but at the same time I feel it is far worse to have every
single module place their `vtable` implementations in slightly different
files (it becomes _incredibly_ hard to find where things are, even with
a language server).

If we really care that the `ArrayVTable` and the `OperationsVTable`
should be located in the same file, then I would argue that those traits
should be combined.

I also snuck in a change (first commit) where we validate in
`new_unchecked` in debug mode only.

Note: All modules have files moved around in the exact same way with the
exception of `primitive`, where I create an `array` module instead of an
`array.rs` because there are so many extra things we implement
specifically for `PrimitiveArray`.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>