Feature: Add Lance to our random access benchmarks (#4882)
Seems like adding lance adds a lot of dependencies. (Edit: for lance we
make an exception and add it only to the `bench-vortex` dependencies,
not the workspace)

Also, it looks like they are also on DF v50 which is nice!

Lance beats us on random access by about ~15% on average (on my
macbook), which isn't too surprising. But also this specific benchmark
is so fast and both are well over 100x faster than Parquet.

In the next few days I will do my best to add lance to all of those
other benchmarks that are `unimplemented!`.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>