Chore: feature flag lance (#4938)
Also modifies the actions so that we only run lance benchmarks in 2
places: _after_ a PR is merged and during nightly benchmarks.

Also let's hold off on merging this until tomorrow just in case it
breaks something on the website.

Also note that the benchmark summaries on github (below) are going to be
off since they no longer include lance

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>