Feature: Make `ListView` the canonical encoding for `List` (#4946)
Tracking Issue: https://github.com/vortex-data/vortex/issues/4699

Revived from: https://github.com/vortex-data/vortex/pull/4853 and
https://github.com/vortex-data/vortex/pull/4830

This PR makes `ListViewArray` the canonical encoding of `DType::List`,
which used to be `ListArray`.

We will probably have a large regression in term of performance for now,
but that can be improved in the future.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>