performance[vortex-datafusion]: support pushing down filters on struct columns (#5024)
The motivation for this PR is to allow pushing down filters of the type
`my_struct.field = 'foo'`. There are some related fixes for things that
broke when I added support for these, namely rewriting the filter on a
physical struct schema that was a subset of the full table schema and
right-hand side literals that are of the dictionary type. Please see
commits for details.

---------

Signed-off-by: Alfonso Subiotto Marques <alfonso.subiotto@polarsignals.com>