Feature: Add methods to `PVectorMut` similar to `std::vec::Vec` (#5079)
Adds methods to `PVec` that are similar in nature to the methods
available on the standard library
[`Vec`](https://doc.rust-lang.org/std/vec/struct.Vec.html#)
implementation.

Note that none of these are strictly necessary right now, but I wanted
to have this now so we can write more meaningful tests (instead of ones
that just track metadata like length and capacity).

I want to add these methods in the future, but they should probably come
in a separate PR:

- `into_iter`
- `retain` (that takes a `Fn(Option<&T>) -> bool` instead of `Fn(&T) ->
bool`
- `retain_mut` (similar to above but `&mut T`
- `retain_indices` (we can use this to implement filter for primitive
arrays)
- `extend_from_slice` (a nice to have because we know the exact length
of the slice and don't have to do all the weird iterator size hint
stuff)
- `copy_within`
- `fill_range`
- `drain` (maybe we don't need this but I can see it being useful)
- `splice` (same)
- `remove`
- `swap_remove`

There are others, but probably a lot less important

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>