Feature: Add `IntoIter` implementation for vectors (#5093)
Adds `IntoIter` impls for primitive and bool vector.

Consolidates all the extend, from_iter, and (new) into_iter stuff into
`iter.rs` files.

The non-cosmetic changes are at the bottom of the `iter.rs` files.

There is a world where we could use the internal `BufferIterator` types
but this was way simpler.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>