Chore: Add `match_vector_pair` macro (#5123)
Note that the formatting is ugly because of this:
https://github.com/rust-lang/rust/pull/52234#issuecomment-903419099

Several of the match statements keep getting out of sync because we have
to remember to add variants to match statements everywhere.

Also fixes some of the macro imports since we're `macro_export`ing
everything again.

Also fixes an issue with the struct API (accidentally didn't return an
`Arc`)

(Note that this is quite important for the implementation of
`FixedSizeList` and `ListView` vectors) otherwise there is way too much
boilerplate

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>