Vortex Session (#5111)
# VortexSession

We have had a `VortexSession` object for a while, but never really made
use of it, in a large part because none of the Vortex components could
take a dependency without creating a cycle.

Since Vortex is incredibly extensible, the idea of the session is to
hold the plugin registries for arrays, layouts, expressions and so on.
While our existing solution does this, it created very fragile APIs
where it was possible to construct file readers/writers _without_ using
the main registry of plugins. Similarly, metrics registries were getting
lost and some metrics ended up in isolated `VortexMetrics::default()`
instances - never to see the light of day.

This PR introduces a top-level VortexSession create that essentially
acts as a type map. Components of Vortex can register session state, and
then read it back again later. This allows components to accept and hold
a VortexSession internally, while our root `vortex` crate performs a
default configuration.

# Runtime Handles

One of the most fragile bits of mis-configured API was the Vortex
runtime handle. This is our abstraction over runtimes (in practice, only
over Tokio or a custom smol-based CurrentThreadRuntime) and provides an
API for components to spawn CPU, I/O, blocking and other types of work.

This PR demonstrates the session API by holding a runtime handle and
plumbing this through APIs that formerly accepted a with_handle
argument.

# Language Bindings

This PR doesn't not completely migrate language bindings onto a
session-based API. Many bindings still use a global static session where
a session instance would be preferred.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>