Performance: `ListViewArray` zero-copyable to `ListArray` optimizations (#5129)
Tracking Issue: https://github.com/vortex-data/vortex/issues/4699
Closes: https://github.com/vortex-data/vortex/issues/5053

This PR introduces a performance optimization for `ListViewArray` by
adding an `is_zero_copy_to_list` flag that tracks whether the
`ListViewArray` can be (near) constant-time converted to a `ListArray`
without copying data.

When this flag is true (indicating sorted offsets with no gaps and no
overlaps), conversions can bypass the very expensive rebuild process
(which just calls `append_scalar` in a loop).

Additionally, this PR refactors all `ListViewArray` constructor call
sites and compute operations throughout the codebase to properly
maintain this new invariant. The only time you can set this flag is via
the `new_unchecked` constructor which is unsafe, which means the caller
better be careful to know what they are doing.

Some things that are not included in this PR:
- Making conversion _actually_ zero-copy would require implementing
logic to always allocate an additional `offsets` slot for an `n+1`th
offset in `ListViewArray`. This is because `ListArray` has `n+1`
offsets, and from a plain `ListViewArray` we have to reallocate the
offsets to make space.
- `reset_offsets` right now is probably more expensive than it could be
since it uses the `sub_scalar` compute function. We know in advance that
`reset_offsets` should _never_ fail in the non-recursive case.
- Validation of this `is_zctl` flag is **SUPER** expensive, but it can
be made a bit less expensive by using a `BitBuffer` to track references
+ a flag for overlaps.
- It would probably be nice to have an `check_is_zero_copy_to_list`
function on `ListViewArray` that sets the flag if it detects it is
zero-copyable to a `ListArray`. This would be nice when we are reading
from disk or converting from an Arrow `GenericListView`.
- This PR **does not** change the on-disk format, which means we lose
this metadata when writing to disk. I think that this flag is somewhat
similar to the `is_sorted` statistic, but at the same time I feel like
it is not correct to store this as a statistic? Not sure.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>