fix[vortex-dict]: avoid full materialization in min_max
Prior to this commit, min_max on dictionaries would materialize the full
dictionary. This commit iterates over the codes and checks whether they fully
cover the values slice, in which case min_max can be run on the values slice.
Even if the codes do not fully cover the values, we can materialize without
duplicates, by filtering the values slice to only the covered indices.

Signed-off-by: Alfonso Subiotto Marques <alfonso.subiotto@polarsignals.com>