vortex-datafusion: Only use schema mapping for casting
When the table schema and the file schema have the same columns but a
different order, then the projection already reorders them and
previously the schema mapper attempted to also fix the ordering, causing
an incorrect result. With this patch, the schema mapper is only used to
perform the expected casting, while leaving the potential reordering to
the projection.

Signed-off-by: Frederic Branczyk <fbranczyk@gmail.com>