Expression execution to be passed input vectors (#5507)
This execution model trades off short-circuiting in favor of CSE.

It also allows expressions to be executed independently of the
expression tree, which may be useful depending on how we ultimately end
up wrapping up scalar expressions into arrays.

@joseph-isaacs the alternative would be to pass a `trait ExecutionArgs`
and then calling `child(usize)` could lazily compute the child, bringing
back the flexibility of short-circuiting depending on how the expression
is ultimately executed.

Signed-off-by: Nicholas Gates <nick@nickgates.com>